spring-core/step04/src/main/java/me/zw/step04/order/OrderServiceImpl.java
package me.zw.step04.order;

import lombok.RequiredArgsConstructor;
import me.zw.step04.annotation.MainDiscountPolicy;
import me.zw.step04.discount.DiscountPolicy;
import me.zw.step04.member.Member;
import me.zw.step04.member.MemberRepository;
import org.springframework.stereotype.Component;

@Component
public class OrderServiceImpl implements OrderService {

    private final MemberRepository memberRepository;

    private final DiscountPolicy discountPolicy;

    public OrderServiceImpl(MemberRepository memberRepository, @MainDiscountPolicy DiscountPolicy discountPolicy) {
        this.memberRepository = memberRepository;
        this.discountPolicy = discountPolicy;
    }

    @Override
    public Order createOrder(Long memberId, String itemName, int itemPrice) {
        Member member = memberRepository.findById(memberId);
        int discountPrice = discountPolicy.discount(member, itemPrice);

        return new Order(memberId, itemName, itemPrice, discountPrice);
    }
}
