spring-core/step04/src/test/java/me/zw/step04/order/OrderServiceImplTest.java
package me.zw.step04.order;

import me.zw.step04.discount.FixedDiscountPolicy;
import me.zw.step04.member.Grade;
import me.zw.step04.member.Member;
import me.zw.step04.member.MemoryMemberRepository;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class OrderServiceImplTest {

    @Test
    void createOrder() {
        MemoryMemberRepository memoryMemberRepository = new MemoryMemberRepository();
        memoryMemberRepository.save(new Member(1L, "name", Grade.VIP));

        OrderServiceImpl orderService = new OrderServiceImpl(memoryMemberRepository, new FixedDiscountPolicy());

        Order order = orderService.createOrder(1L, "itemA", 10000);
        assertThat(order.getDiscountPrice()).isEqualTo(1000);
    }

}